Se cambió el tipado de la function authenticate, se cambió el nombre del archivo AuthStore a useAuthStore y se cambio el uso de Staff a Nurse, creando también un tipado más correcto para la const nurse
